+++ src/client/js/controllers/main.js
@@ -147,16 +147,17 @@ function MainController($rootScope, $scope, $state, $http, commonFactory) {
   }
 
   $scope.deleteSurvey = function(id) {
-    $http.delete("/api/surveys/" + id)
-      .then(function(response) {
-        // success callback
-        commonFactory.activateAlert('Encuesta fue borrada exitosamente!', 'info');
-        retrieveSurveys();
-      })
-      .catch(function(error) {
-        commonFactory.activateAlert('Woops! Algo paso!', 'danger');
-      });
-
+    if (confirm("Esta seguro de borrar esta encuesta?")) {
+      $http.delete("/api/surveys/" + id)
+        .then(function(response) {
+          // success callback
+          commonFactory.activateAlert('Encuesta fue borrada exitosamente!', 'info');
+          retrieveSurveys();
+        })
+        .catch(function(error) {
+          commonFactory.activateAlert('Woops! Algo paso!', 'danger');
+        });
+    }
   }
 
   function initializeController() {
